# docs/api/openapi_v2.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RuntimeConfigResponseV2": {
        "description": "Typed, versioned envelope around the runtime configuration document.",
        "properties": {
          "config": {
            "$ref": "#/components/schemas/RuntimeConfig"
          },
          "instructions": {
            "$ref": "#/components/schemas/RuntimeInstructions"
          },
          "schemaVersion": {
            "default": "2.0",
            "title": "Schemaversion",
            "type": "string"
          }
        },
        "required": [
          "config",
          "instructions"
        ],
        "title": "RuntimeConfigResponseV2",
        "type": "object"
      },
...
      "RuntimeConfigUpdateRequestV2": {
        "description": "Request body for ``PUT /v2/config``.",
        "properties": {
          "config": {
            "$ref": "#/components/schemas/RuntimeConfig"
          }
        },
        "required": [
          "config"
        ],
        "title": "RuntimeConfigUpdateRequestV2",
        "type": "object"
      },
...
      "SessionListV2": {
        "description": "Paginated collection of :class:`SessionV2`.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/SessionV2"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "$ref": "#/components/schemas/PageMetaV2"
          },
          "schemaVersion": {
            "default": "2.0",
            "title": "Schemaversion",
            "type": "string"
          }
        },
        "required": [
          "items",
          "page"
        ],
        "title": "SessionListV2",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "oidcBearer": {
        "bearerFormat": "JWT",
        "description": "OIDC-issued bearer JWT, validated against the configured JWKS.",
        "scheme": "bearer",
        "type": "http"
      },
      "serviceBearer": {
        "bearerFormat": "adk_live_<key-id>_<secret>",
        "description": "Governed service credential minted via `autodev auth service-key create`.",
        "scheme": "bearer",
        "type": "http"
      },
      "sessionCookie": {
        "description": "Browser session established by the OIDC PKCE login flow.",
        "in": "cookie",
        "name": "autodev_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "AutoDev Orchestrator - Control Plane API /v2",
    "version": "0.3.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v2/config": {
      "get": {
        "description": "Return the current runtime configuration with secrets redacted.\n\nArgs:\n    config_service: Shared runtime config service (same singleton v1's\n        ``GET /config`` uses).\n\nReturns:\n    The versioned configuration document.",
        "operationId": "get_runtime_config_v2_v2_config_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeConfigResponseV2"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "oidcBearer": []
          },
          {
            "serviceBearer": []
          },
          {
            "sessionCookie": []
          }
        ],
        "summary": "Get Runtime Config V2",
        "tags": [
          "config"
        ],
        "x-autodev-required-scope": "config:read_redacted"
      },
      "put": {
        "description": "Persist a new runtime configuration and apply it to the process.\n\nArgs:\n    request: The new configuration to persist.\n    config_service: Shared runtime config service (same singleton v1's\n        ``PUT /config`` uses).\n\nReturns:\n    The versioned configuration document after the update.",
        "operationId": "update_runtime_config_v2_v2_config_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuntimeConfigUpdateRequestV2"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeConfigResponseV2"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "oidcBearer": []
          },
          {
            "serviceBearer": []
          },
          {
            "sessionCookie": []
          }
        ],
        "summary": "Update Runtime Config V2",
        "tags": [
          "config"
        ],
        "x-autodev-required-scope": "config:write_safe"
      }
    },
...
    "/v2/sessions": {
      "get": {
        "description": "List all known sessions.\n\nArgs:\n    pagination: Shared limit/offset pagination window.\n    orchestrator: Orchestrator service dependency.\n\nReturns:\n    A paginated collection of sessions.",
        "operationId": "list_sessions_v2_v2_sessions_get",
        "parameters": [
          {
            "description": "Max items to return.",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "description": "Max items to return.",
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Number of items to skip.",
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "description": "Number of items to skip.",
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionListV2"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "oidcBearer": []
          },
          {
            "serviceBearer": []
          },
          {
            "sessionCookie": []
          }
        ],
        "summary": "List Sessions V2",
        "tags": [
          "sessions"
        ],
        "x-autodev-required-scope": "session:read"
      },
      "post": {
        "description": "Create a new session and generate its initial plan.\n\nArgs:\n    request: The session creation request (goal).\n    orchestrator: Orchestrator service dependency.\n\nReturns:\n    The newly created session.",
        "operationId": "create_session_v2_v2_sessions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionCreateRequestV2"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionV2"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "oidcBearer": []
          },
          {
            "serviceBearer": []
          },
          {
            "sessionCookie": []
          }
        ],
        "summary": "Create Session V2",
        "tags": [
          "sessions"
        ],
        "x-autodev-required-scope": "session:write"
      }
    },
    "/v2/sessions/{session_id}": {
      "get": {
        "description": "Fetch a single session by id.\n\nArgs:\n    session_id: Identifier of the session.\n    orchestrator: Orchestrator service dependency.\n\nReturns:\n    The requested session.\n\nRaises:\n    HTTPException: 404 if ``session_id`` does not exist.",
        "operationId": "get_session_v2_v2_sessions__session_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionV2"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "oidcBearer": []
          },
          {
            "serviceBearer": []
          },
          {
            "sessionCookie": []
          }
        ],
        "summary": "Get Session V2",
        "tags": [
          "sessions"
        ],
        "x-autodev-required-scope": "session:read"
      }
    },
    "/v2/sessions/{session_id}/execution-plan": {
      "get": {
        "description": "Derive an execution plan from a session's accumulated agent artifacts.\n\nArgs:\n    session_id: Identifier of the session.\n    orchestrator: Orchestrator service dependency.\n\nReturns:\n    The derived execution plan.\n\nRaises:\n    HTTPException: 404 if ``session_id`` does not exist.",
        "operationId": "get_execution_plan_v2_v2_sessions__session_id__execution_plan_get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecutionPlanV2"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "oidcBearer": []
          },
          {
            "serviceBearer": []
          },
          {
            "sessionCookie": []
          }
        ],
        "summary": "Get Execution Plan V2",
        "tags": [
          "planning"
        ],
        "x-autodev-required-scope": "session:read"
      }
    },
...
  }
}